requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.7
django-toolbelt==0.0.1
djangorestframework==3.12.2
djangorestframework-gis==0.17
djangorestframework-simplejwt==4.6.0
geojson==2.5.0
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.6
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.4
